CI: Run workflow on push to any branch
There is no need to restrict our workflow to only run after a push to
the main branch. This should increase flexibility during development to
run the CI workflow on other branches. GitHub does not limit the time
available for CI runs for public repositories, so there is not any
downside to this change.